renamed (Const)Iterator to (const_)iterator; changed size() to return an uint
svn-id: r12722

+++ gui/EditTextWidget.cpp
@@ -59,7 +59,7 @@ void EditTextWidget::handleMouseDown(int x, int y, int button, int clickCount){
 	x += _labelOffset;
 
 	int width = 0;
-	int i;
+	uint i;
 
 	for (i = 0; i < _label.size(); ++i) {
 		width += gui->getCharWidth(_label[i]);
@@ -112,7 +112,7 @@ bool EditTextWidget::handleKeyDown(uint16 ascii, int keycode, int modifiers) {
 		}
 		break;
 	case 256 + 19:	// right arrow
-		if (_pos < _label.size()) {
+		if (_pos < (int)_label.size()) {
 			_pos++;
 			dirty = adjustOffset();
 		}
